year summary, currency to big decimal

+++ src/main/kotlin/com/example/demo/entity/ShoppingItem.kt
@@ -1,4 +1,6 @@
 package com.example.demo.entity
 
-class ShoppingItem(val id: Long, val productName: String, val quantity: Double,
-                   val price: Double, val unitPrice: Double, val description: String = "")
+import java.math.BigDecimal
+
+class ShoppingItem(val id: Long, val productName: String, val quantity: BigDecimal,
+                   val price: BigDecimal, val unitPrice: BigDecimal, val description: String = "")
